call graph actions moved to their own service
