+++ resharper/resharper-unity/src/CSharp/Feature/Services/CallGraph/BurstCodeAnalysis/AddDiscardAttribute/AddDiscardAttributeBulbAction.cs
@@ -7,7 +7,7 @@ using JetBrains.ReSharper.Psi.CSharp.Tree;
 using JetBrains.ReSharper.Resources.Shell;
 using JetBrains.TextControl;
 
-namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.QuickFixes.CallGraph.BurstCodeAnalysis.
+namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.CallGraph.BurstCodeAnalysis.
     AddDiscardAttribute
 {
     public sealed class AddDiscardAttributeBulbAction : IBulbAction
