+++ resharper/resharper-unity/src/CSharp/Feature/Services/CallGraph/BurstCodeAnalysis/AddDiscardAttribute/AddDiscardAttributeContextAction.cs
@@ -7,7 +7,7 @@ using JetBrains.ReSharper.Feature.Services.Intentions;
 using JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.ContextActions;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
 
-namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.QuickFixes.CallGraph.BurstCodeAnalysis.
+namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.CallGraph.BurstCodeAnalysis.
     AddDiscardAttribute
 {
     [ContextAction(
