+++ resharper/resharper-unity/src/CSharp/Feature/Services/CallGraph/BurstCodeAnalysis/AddDiscardAttribute/AddDiscardAttributeUtil.cs
@@ -1,4 +1,4 @@
-namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.QuickFixes.CallGraph.BurstCodeAnalysis.AddDiscardAttribute
+namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.CallGraph.BurstCodeAnalysis.AddDiscardAttribute
 {
     public static class AddDiscardAttributeUtil
     {
