+++ resharper/resharper-unity/src/CSharp/Feature/Services/CallGraph/ExpensiveCodeAnalysis/ExpensiveCodeActionsUtil.cs
@@ -1,4 +1,4 @@
-namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.QuickFixes.CallGraph.ExpensiveCodeAnalysis
+namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.CallGraph.ExpensiveCodeAnalysis
 {
     public static class ExpensiveCodeActionsUtil
     {
