+++ resharper/resharper-unity/src/CSharp/Feature/Services/CallGraph/PerformanceAnalysis/PerformanceAnalysisDisableByCommentQuickFix.cs
@@ -4,23 +4,24 @@ using JetBrains.Application.Threading;
 using JetBrains.ReSharper.Feature.Services.Bulbs;
 using JetBrains.ReSharper.Feature.Services.Intentions;
 using JetBrains.ReSharper.Feature.Services.QuickFixes;
-using JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Stages.BurstCodeAnalysis.Highlightings;
+using JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Errors;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
 using JetBrains.Util;
 
-namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.QuickFixes.CallGraph.BurstCodeAnalysis.AddDiscardAttribute
+namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.CallGraph.PerformanceAnalysis
 {
     [QuickFix]
-    public sealed class AddDiscardAttributeQuickFix : IQuickFix
+    public sealed class PerformanceAnalysisDisableByCommentQuickFix : IQuickFix
     {
         [CanBeNull] private readonly IMethodDeclaration myMethodDeclaration;
-        [CanBeNull] private readonly AddDiscardAttributeBulbAction myBulbAction;
-        public AddDiscardAttributeQuickFix(IBurstHighlighting burstHighlighting)
-        {
-            myMethodDeclaration = burstHighlighting?.Node?.GetContainingNode<IMethodDeclaration>();
+        [CanBeNull] private readonly PerformanceAnalysisDisableByCommentBulbAction myBulbAction;
 
+        public PerformanceAnalysisDisableByCommentQuickFix(UnityPerformanceInvocationWarning performanceHighlighting)
+        {
+            myMethodDeclaration = performanceHighlighting.InvocationExpression?.GetContainingNode<IMethodDeclaration>();
+            
             if (myMethodDeclaration != null)
-                myBulbAction = new AddDiscardAttributeBulbAction(myMethodDeclaration);
+                myBulbAction = new PerformanceAnalysisDisableByCommentBulbAction(myMethodDeclaration);
         }
 
         public IEnumerable<IntentionAction> CreateBulbItems()
@@ -29,13 +30,13 @@ namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.QuickFixes.C
                 return EmptyList<IntentionAction>.Instance;
             
             myMethodDeclaration.GetPsiServices().Locks.AssertReadAccessAllowed();
-            
+
             return myBulbAction.ToQuickFixIntentions();
         }
-        
+
         public bool IsAvailable(IUserDataHolder cache)
         {
-            return BurstActionsUtil.IsAvailable(myMethodDeclaration);
+            return PerformanceDisableUtil.IsAvailable(myMethodDeclaration);
         }
     }
 }
