+++ resharper/resharper-unity/src/CSharp/Feature/Services/CallGraph/PerformanceAnalysis/PerformanceDisableUtil.cs
@@ -1,24 +1,26 @@
 using JetBrains.Annotations;
 using JetBrains.Application.Threading;
-using JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Stages.BurstCodeAnalysis;
+using JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Stages.PerformanceCriticalCodeAnalysis;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
 
-namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.QuickFixes.CallGraph.BurstCodeAnalysis
+namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.CallGraph.PerformanceAnalysis
 {
-    public static class BurstActionsUtil
+    public static class PerformanceDisableUtil
     {
+        public const string MESSAGE = "Disable performance analysis for method";
+        
         [ContractAnnotation("null => false")]
         public static bool IsAvailable([CanBeNull] IMethodDeclaration methodDeclaration)
         {
             if (methodDeclaration == null)
                 return false;
-
+            
             methodDeclaration.GetPsiServices().Locks.AssertReadAccessAllowed();
 
             var declaredElement = methodDeclaration.DeclaredElement;
 
             return declaredElement != null && methodDeclaration.IsValid() &&
-                   !BurstCodeAnalysisUtil.IsBurstProhibitedFunction(declaredElement);
+                   !PerformanceCriticalCodeStageUtil.IsPerformanceCriticalRootMethod(methodDeclaration);
         }
     }
 }
