+++ resharper/resharper-unity/test/src/CSharp/Intentions/QuickFixes/PerformanceAnalysisDisableQuickFixTests.cs
@@ -1,7 +1,7 @@
 using JetBrains.Application.Settings;
 using JetBrains.ReSharper.Feature.Services.Daemon;
 using JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Errors;
-using JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.QuickFixes.CallGraph.PerformanceAnalysis;
+using JetBrains.ReSharper.Plugins.Unity.CSharp.Feature.Services.CallGraph.PerformanceAnalysis;
 using JetBrains.ReSharper.Psi;
 using NUnit.Framework;
